Mod: 1
In this modification I have drawn the mouth of the Player just to add a bit more detail.
Also I have made size of all the elements of our game bigger.

This part has not been covered in the tutorial as this is trial and error.

Feel free to download the code